server/workers/requirements.txt
certifi
redis
hiredis
aioredis
pandas
pandas_schema
scikit-learn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask
flask-cors
flask_sqlalchemy
flask_restx
Werkzeug
marshmallow
gunicorn
flasgger
elasticsearch
elasticsearch_dsl
pyyaml
pytest
nltk
spacy
spacy_cld
aiohttp